+++ logic/Actions/ListReadingTypes.php
@@ -3,6 +3,7 @@
 namespace AirQuality\Actions;
 
 use \SBRL\TomlConfig;
+use \SBRL\ResponseEncoder;
 use \AirQuality\Repositories\IMeasurementTypeRepository;
 use \AirQuality\ApiResponseSender;
 
@@ -33,6 +34,13 @@ class ListReadingTypes implements IAction {
 		
 		// 1: Parse & validate parameters
 		$device_id = !empty($_GET["device-id"]) ? intval($_GET["device-id"]) : null;
+		$format = $_GET["format"] ?? "json";
+		if(!in_array($format, ["json", "csv"])) {
+			$this->sender->send_error_plain(406,
+				"Error: The format '$format' isn't recognised. Valid formats: " . implode(", ", $format) . "."
+			);
+			exit;
+		}
 		
 		// 1: Pull data from database
 		$data = null;
@@ -52,7 +60,19 @@ class ListReadingTypes implements IAction {
 		
 		// 3: Serialise data
 		$start_encode = microtime(true);
-		$response = json_encode($data);
+		$response = null;
+		$response_type = "application/octet-stream";
+		$response_suggested_filename = "data-" . date(\DateTime::ATOM) . "";
+		switch($format) {
+			case "json":
+				$response_type = "application/json";
+				$response = json_encode($data);
+				break;
+			case "csv":
+				$response_type = "text/csv";
+				$response = ResponseEncoder::encode_csv($data);
+				break;
+		}
 		
 		// 4: Send response
 		
@@ -60,7 +80,8 @@ class ListReadingTypes implements IAction {
 		// TODO: Investigate adding a short-term (~10mins?) cache-control header here
 		
 		header("content-length: " . strlen($response));
-		header("content-type: application/json");
+		header("content-type: $response_type");
+		header("content-disposition: inline; filename=$response_suggested_filename");
 		header("x-time-taken: " . PerfFormatter::format_perf_data($start_time, $start_handle, $start_encode));
 		echo($response);
 		return true;