+++ Changelog.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## v0.4 - 24th February 2019 (unreleased)
- - [API] Added new `format` GET parameter to `device-data` action.
+ - [API] Added new `format` GET parameter to the following actions:
+     - `device-data`
+     - `list-reading-types`
 
 ## v0.3.3 - 20th February 2019
  - Updated to use new database structure